project hygiene
