+++ targets.props
@@ -32,6 +32,7 @@
         <AppTargetFramework>net462</AppTargetFramework>
         <TestsTargetFrameworks>net462</TestsTargetFrameworks>
         <LibTargetFrameworks>net462</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net462</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     <When  Condition="'$(CustomTestTarget)' == 'netstandard2.0'">
@@ -43,6 +44,7 @@
         <AppTargetFramework>net8.0</AppTargetFramework>
         <TestsTargetFrameworks>net8.0</TestsTargetFrameworks>
         <LibTargetFrameworks>netstandard2.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>netstandard2.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     <When  Condition="'$(CustomTestTarget)' == 'netstandard2.1'">
@@ -54,6 +56,19 @@
         <AppTargetFramework>net8.0</AppTargetFramework>
         <TestsTargetFrameworks>net8.0</TestsTargetFrameworks>
         <LibTargetFrameworks>netstandard2.1</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>netstandard2.1</CryptolibTargetFrameworks>
+      </PropertyGroup>
+    </When>
+    <When  Condition="'$(CustomTestTarget)' == 'net472'">
+      <PropertyGroup>
+        <AnalysisLevel>preview</AnalysisLevel>
+        <AnalysisMode>all</AnalysisMode>
+        <AnalysisModeStyle>default</AnalysisModeStyle>
+        <AppTargetFrameworks>net472</AppTargetFrameworks>
+        <AppTargetFramework>net472</AppTargetFramework>
+        <TestsTargetFrameworks>net472</TestsTargetFrameworks>
+        <LibTargetFrameworks>net472</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net472</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     <When  Condition="'$(CustomTestTarget)' == 'net48'">
@@ -65,6 +80,19 @@
         <AppTargetFramework>net48</AppTargetFramework>
         <TestsTargetFrameworks>net48</TestsTargetFrameworks>
         <LibTargetFrameworks>net48</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net48</CryptolibTargetFrameworks>
+      </PropertyGroup>
+    </When>
+    <When  Condition="'$(CustomTestTarget)' == 'net6.0'">
+      <PropertyGroup>
+        <AnalysisLevel>preview</AnalysisLevel>
+        <AnalysisMode>all</AnalysisMode>
+        <AnalysisModeStyle>default</AnalysisModeStyle>
+        <AppTargetFrameworks>net6.0</AppTargetFrameworks>
+        <AppTargetFramework>net6.0</AppTargetFramework>
+        <TestsTargetFrameworks>net6.0</TestsTargetFrameworks>
+        <LibTargetFrameworks>net6.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net6.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     <When  Condition="'$(CustomTestTarget)' == 'net8.0'">
@@ -76,6 +104,7 @@
         <AppTargetFramework>net8.0</AppTargetFramework>
         <TestsTargetFrameworks>net8.0</TestsTargetFrameworks>
         <LibTargetFrameworks>net8.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net8.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     <When  Condition="'$(CustomTestTarget)' == 'net9.0'">
@@ -87,6 +116,7 @@
         <AppTargetFramework>net9.0</AppTargetFramework>
         <TestsTargetFrameworks>net9.0</TestsTargetFrameworks>
         <LibTargetFrameworks>net9.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net9.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     <When  Condition="'$(CustomTestTarget)' == 'net10.0'">
@@ -98,6 +128,7 @@
         <AppTargetFramework>net10.0</AppTargetFramework>
         <TestsTargetFrameworks>net10.0</TestsTargetFrameworks>
         <LibTargetFrameworks>net10.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net10.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
 
@@ -111,6 +142,7 @@
         <AppTargetFramework>net10.0</AppTargetFramework>
         <TestsTargetFrameworks>net10.0</TestsTargetFrameworks>
         <LibTargetFrameworks>net10.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net10.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     
@@ -124,6 +156,7 @@
         <AppTargetFramework>net10.0</AppTargetFramework>
         <TestsTargetFrameworks>net48;net8.0;net10.0</TestsTargetFrameworks>
         <LibTargetFrameworks>net462;netstandard2.0;netstandard2.1;net8.0;net10.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net462;net472;netstandard2.0;netstandard2.1;net8.0;net10.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     
@@ -137,6 +170,7 @@
         <AppTargetFramework>net8.0</AppTargetFramework>
         <TestsTargetFrameworks>net48;net8.0;net9.0</TestsTargetFrameworks>
         <LibTargetFrameworks>net462;netstandard2.0;netstandard2.1;net8.0;net9.0</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net462;net472;netstandard2.0;netstandard2.1;net8.0;net9.0</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     
@@ -148,6 +182,7 @@
         <AppTargetFramework>net48</AppTargetFramework>
         <TestsTargetFrameworks>net48</TestsTargetFrameworks>
         <LibTargetFrameworks>net48;netstandard2.1</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net48;netstandard2.1</CryptolibTargetFrameworks>
       </PropertyGroup>
     </When>
     
@@ -158,6 +193,7 @@
         <AppTargetFramework>net48</AppTargetFramework>
         <TestsTargetFrameworks>net48</TestsTargetFrameworks>
         <LibTargetFrameworks>net48</LibTargetFrameworks>
+        <CryptolibTargetFrameworks>net48</CryptolibTargetFrameworks>
       </PropertyGroup>
     </Otherwise>
   </Choose>
