+++ tests/Security/Cryptography/Hash/Keccak/KT256Tests.cs
@@ -18,17 +18,6 @@ using System.Text;
 [Parallelizable(ParallelScope.All)]
 public class KT256Tests
 {
-    private static byte[] HexToBytes(string hex)
-    {
-        hex = hex.Replace(" ", "");
-        byte[] result = new byte[hex.Length / 2];
-        for (int i = 0; i < result.Length; i++)
-        {
-            result[i] = Convert.ToByte(hex.Substring(i * 2, 2), 16);
-        }
-        return result;
-    }
-
     private static byte[] CreatePattern(int length)
     {
         byte[] result = new byte[length];
@@ -45,7 +34,7 @@ public class KT256Tests
     [Test]
     public void EmptyInputEmptyCustomization64Bytes()
     {
-        byte[] expected = HexToBytes(
+        byte[] expected = TestHelpers.FromHexString(
             "B2 3D 2E 9C EA 9F 49 04 E0 2B EC 06 81 7F C1 0C E3 8C E8 E9 3E F4 C8 9E 65 37 07 6A F8 64 64 04 " +
             "E3 E8 B6 81 07 B8 83 3A 5D 30 49 0A A3 34 82 35 3F D4 AD C7 14 8E CB 78 28 55 00 3A AE BD E4 A9");
 
@@ -61,7 +50,7 @@ public class KT256Tests
     [Test]
     public void Pattern1ByteEmptyCustomization()
     {
-        byte[] expected = HexToBytes(
+        byte[] expected = TestHelpers.FromHexString(
             "0D 00 5A 19 40 85 36 02 17 12 8C F1 7F 91 E1 F7 13 14 EF A5 56 45 39 D4 44 91 2E 34 37 EF A1 7F " +
             "82 DB 6F 6F FE 76 E7 81 EA A0 68 BC E0 1F 2B BF 81 EA CB 98 3D 72 30 F2 FB 02 83 4A 21 B1 DD D0");
         byte[] input = CreatePattern(1);
@@ -78,7 +67,7 @@ public class KT256Tests
     [Test]
     public void Pattern17BytesEmptyCustomization()
     {
-        byte[] expected = HexToBytes(
+        byte[] expected = TestHelpers.FromHexString(
             "1B A3 C0 2B 1F C5 14 47 4F 06 C8 97 99 78 A9 05 6C 84 83 F4 A1 B6 3D 0D CC EF E3 A2 8A 2F 32 3E " +
             "1C DC CA 40 EB F0 06 AC 76 EF 03 97 15 23 46 83 7B 12 77 D3 E7 FA A9 C9 65 3B 19 07 50 98 52 7B");
         byte[] input = CreatePattern(17);
